% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{backup_rs_snippets}
\alias{backup_rs_snippets}
\title{Create a backup of snippets file}
\usage{
backup_rs_snippets(type)
}
\arguments{
\item{type}{One of \code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"},
\code{"java"}, \code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
Default is \code{"r"}.}
}
\description{
Create a backup of snippets file
}
