% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets--backup.R
\name{backup_rstudio_snippets}
\alias{backup_rstudio_snippets}
\alias{list_snippet_file_backups}
\alias{open_snippets_backup_dir}
\alias{remove_snippet_backup_duplicates}
\title{Manage back-up copies of snippets}
\usage{
backup_rstudio_snippets(type)

list_snippet_file_backups(type = get_default_snippet_types())

open_snippets_backup_dir()

remove_snippet_backup_duplicates()
}
\arguments{
\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}
}
\description{
\itemize{
\item \code{backup_rstudio_snippets()} creates a back-up of snippets file.
\item \code{list_snippet_file_backups()} lists the names of current file with snippets
and its back-ups.
\item \code{remove_snippet_backup_duplicates()} removes duplicated backup files.
}
}
\examples{
if (FALSE) {

backup_rstudio_snippets("r")
backup_rstudio_snippets("markdown")

list_snippet_file_backups("r")

}
}
\concept{backup snippets}
