% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{banner}
\alias{banner}
\title{banner}
\usage{
banner(nams, values, len = 80, sql = FALSE)
}
\description{
banner
}
\examples{
values = c('1', 'Aname Bname', paste(sample(iris$Species, 20), collapse = ' ' ) )
nams = c('ID', 'Name', NA)
 cat( banner(nams, values, 50)  )
}
