% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{make_snippet_filename}
\alias{make_snippet_filename}
\title{Make a filename for snippets}
\usage{
make_snippet_filename(type = get_default_snippet_types())
}
\arguments{
\item{type}{A character vector of candidate values.
One of \code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"},
\code{"java"}, \code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
The default is \code{"r"}.}
}
\value{
String with a filename.
}
\description{
Make a filename for certain type of snippets.
}
\examples{
make_snippet_filename()

make_snippet_filename("markdown")

}
