% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet-types.R
\name{match_snippet_type}
\alias{match_snippet_type}
\title{Return correct snippet type}
\usage{
match_snippet_type(type = get_default_snippet_types(), several.ok = FALSE)
}
\arguments{
\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{several.ok}{(logical)
Specify if \code{type} should be allowed to have more than one element.
Defaults to \code{FALSE}.
See also \code{\link[base:match.arg]{base::match.arg()}}}
}
\value{
(sting) Correct snippet type in lower case. By default returns \code{"r"}.
}
\description{
Return correct snippet type
}
\examples{
# Defaults to "r":
match_snippet_type()

match_snippet_type("r")

match_snippet_type("m")
}
\concept{snippet types}
