% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet-files-and-dirs.R
\name{open_rs_snippets_dir}
\alias{open_rs_snippets_dir}
\title{Open directory of RStudio snippets}
\usage{
open_rs_snippets_dir()
}
\description{
Open directory of RStudio snippets
}
\concept{snippet files and dirs}
