% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets--files-and-dirs.R
\name{open_rs_snippets_file}
\alias{open_rs_snippets_file}
\title{Edit RStudio Snippets}
\usage{
open_rs_snippets_file(type, rstudio_version = "auto")
}
\arguments{
\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{rstudio_version}{Numeric version of RStudio (the output is different
for RStudio < 1.3 and 1.3 or newer series). If \code{"auto"}, the version
of RStudio will be determined automatically.}
}
\description{
Edit file with RStudio snippets.
}
\examples{
\dontrun{\donttest{
open_rs_snippets_file("r")

open_rs_snippets_file("markdown")
}}

}
\concept{snippet files and dirs}
