% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet_files_and_dirs.R
\name{path_to_rs_snippets_dir}
\alias{path_to_rs_snippets_dir}
\title{Get path to directory for RStudio snippets}
\usage{
path_to_rs_snippets_dir()
}
\value{
(string) Path to directory for RStudio snippets.
}
\description{
Create a string with a path to directory for RStudio snippets.
The directory might not exist.
}
\examples{
path_to_rs_snippets_dir()
}
\concept{snippet files and dirs}
