% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet-files-and-dirs.R
\name{path_to_snippets_dir_of_pkg}
\alias{path_to_snippets_dir_of_pkg}
\alias{path_to_snippets_files_of_pkg}
\title{Get path to snippets in a package}
\usage{
path_to_snippets_dir_of_pkg(package, ...)

path_to_snippets_files_of_pkg(package, ...)
}
\arguments{
\item{package}{(character) Package name.}

\item{...}{(character) Path to subdirectory with snippets of interest:
\verb{\{path to package\}/inst/snippets/\{path of subdirectory provided via ...\}}}
}
\description{
Get path to snippets in a package:
}
\details{
\itemize{
\item \code{path_to_snippets_dir_of_pkg()} gets path to directory with snippet files
in a package. Defaults to \verb{\{path to package\}/inst/snippets}. Returns empty
string, if the directory does not exist.
\item \code{path_to_snippets_files_of_pkg()} gets paths to all files with snippets
in a package. If the directory of interest does not exist, \code{NULL} is
returned.
}
}
\examples{
path_to_snippets_dir_of_pkg("snippets")

path_to_snippets_files_of_pkg("snippets")
}
