% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet_files_and_dirs.R
\name{path_to_snippets_file}
\alias{path_to_snippets_file}
\alias{path_to_rs_snippets_file}
\title{Construct path to file of certain type of snippets}
\usage{
path_to_snippets_file(
  dir,
  type = get_default_snippet_types(),
  create = FALSE,
  several.ok = FALSE
)

path_to_rs_snippets_file(
  type = get_default_snippet_types(),
  create = FALSE,
  several.ok = FALSE
)
}
\arguments{
\item{dir}{(string) Directory name.}

\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{create}{(logical) If \code{TRUE}, as a side effect, the file is created
(if it does not exist). Defaults to \code{FALSE}.}

\item{several.ok}{(logical)
Specify if \code{type} should be allowed to have more than one element.
Defaults to \code{FALSE}.
See also \code{\link[base:match.arg]{base::match.arg()}}}
}
\value{
(character) Path to file.
}
\description{
Create a string with a path to file of certain type of snippets:
}
\details{
\itemize{
\item \code{path_to_snippets_file()}    in any folder.
\item \code{path_to_rs_snippets_file()} in a folder from which RStudio reads
snippets.
}
}
\concept{snippet files and dirs}
