% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{read_snippets}
\alias{read_snippets}
\title{Read snippets into data frame.}
\usage{
read_snippets(input)
}
\arguments{
\item{input}{(character)
Either a file name or a character vector with with a snippet.}
}
\value{

}
\description{
Read snippets into data frame.
}
\examples{

library(tidyverse)
library(snippets)

file <- system.file("test/html.snippets", package = "snippets")
read_snippets(file)

file2 <- path_to_snippets_files_of_pkg("snippets")[1]
read_snippets(file2)


text <- (
'
snippet space
	&${1:nbsp};${0}

snippet nbsp
	&${1:nbsp};${0}
')

read_snippets(text)
}
