% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{edit_rstudio_snippets}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{usethis}{\code{\link[usethis]{edit_rstudio_snippets}}}
}}

