% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{replace_snippets_file}
\alias{replace_snippets_file}
\title{Replace snippets file}
\usage{
replace_snippets_file(
  type = get_default_snippet_types(),
  from_dir = get_pkg_snippets_dir("custom"),
  backup = TRUE
)
}
\arguments{
\item{type}{The type of RStudio snippet.}

\item{from_dir}{The directory with replacement file.}

\item{backup}{(logical) Indication if a backup copy should be created.}
}
\description{
Replace snippets file
}
\examples{
if (FALSE) {

# Replace your R and Markdown snippets with those in package "snippets":
replace_snippets_file("r",        backup = TRUE)
replace_snippets_file("markdown", backup = TRUE)

# Check if backup copies exist:
list_snippet_file_backups("r")
list_snippet_file_backups("markdown")
}
}
