% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.R
\name{run_snippets}
\alias{run_snippets}
\title{check all snippets (for now R and mysql snippets)}
\usage{
run_snippets()
}
\value{
numeric: the number of updated rows.
}
\description{
check all snippets (for now R and mysql snippets)
}
