% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.R
\name{runSnippet}
\alias{runSnippet}
\title{check one snippet}
\usage{
runSnippet(snippet, lang)
}
\value{
datetime if the snippet runs without errors else NA
}
\description{
check one snippet
}
\examples{
runSnippet(snippet = 'x = rnorm(10)', lang = 'r')
runSnippet(snippet = 'SELECT 1', lang = 'mysql')
}
