% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets-package.R
\docType{package}
\name{snippets-package}
\alias{snippets}
\alias{snippets-package}
\title{snippets: Manage RStudio Snippets}
\description{
Manage RStudio snippets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gegznav.github.io/snippets/}
  \item Report bugs at \url{https://github.com/GegznaV/snippets/issues}
}

}
\author{
\strong{Maintainer}: Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID})

}
\keyword{internal}
