% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{snippets_file_exists}
\alias{snippets_file_exists}
\title{Does file with snippets exist?}
\usage{
snippets_file_exists(type)
}
\arguments{
\item{type}{One of \code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"},
\code{"java"}, \code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
Default is \code{"r"}.}
}
\value{

}
\description{
Does a file with certain types of snippets exist in RStudio snippets directory?
}
\examples{
snippets_file_exists("r")
snippets_file_exists("markdown")
}
