% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{snipSave}
\alias{snipSave}
\alias{snipFetch}
\alias{snipDrop}
\alias{snipSearch}
\title{SQL Snippets}
\usage{
snipSave(snippet, lang, author = NA, description = NA, ID)

snipFetch(ID, verbose = FALSE, banner = TRUE, asis = FALSE)

snipDrop(ID)

snipSearch(kw, lang)
}
\arguments{
\item{snippet}{The snippet}

\item{lang}{language: R, python, bash}

\item{author}{author}

\item{description}{description}

\item{ID}{snip ID - will overwrite that ID}
}
\value{
snipSave returns the ID of the saved snippet 
					   or NULL if this fails.
}
\description{
SQL Snippets
}
