% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet-files-and-dirs.R
\name{create_rs_snippets_dir}
\alias{create_rs_snippets_dir}
\title{Create directory for RStudio snippets}
\usage{
create_rs_snippets_dir()
}
\value{
Invisibly returns the path to created directory.
See also \code{\link[fs:create]{fs::dir_create()}}.
}
\description{
Create directory for RStudio snippets
}
\examples{
if (FALSE) {
create_rs_snippets_dir()
}
}
\concept{snippet files and dirs}
