% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets--snippet-types.R
\name{get_default_snippet_types}
\alias{get_default_snippet_types}
\title{Get default snippet types}
\usage{
get_default_snippet_types()
}
\value{
Character vector with possible snippet type values.
}
\description{
Get default snippet types
}
\examples{
get_default_snippet_types()
#  "r.snippets",
#  "markdown.snippets",
#  "c_cpp.snippets",
#  "css.snippets",
#  "html.snippets",
#  "java.snippets",
#  "javascript.snippets",
#  "python.snippets",
#  "sql.snippets",
#  "stan.snippets",
#  "tex.snippets"
#  stringr::str_extract(".*(?=\.snippets)")
}
\concept{snippet types}
