% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet_files_and_dirs.R
\name{get_path_to_snippet_file}
\alias{get_path_to_snippet_file}
\alias{get_path_to_rs_snippet_file}
\title{Construct path to file of certain type snippets.}
\usage{
get_path_to_snippet_file(
  dir,
  type = get_default_snippet_types(),
  create = FALSE,
  several.ok = FALSE
)

get_path_to_rs_snippet_file(
  type = get_default_snippet_types(),
  create = FALSE,
  several.ok = FALSE
)
}
\arguments{
\item{dir}{(string) Directory name.}

\item{type}{(character)
A character vector of candidate values.
One of \code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{create}{(logical) If \code{TRUE}, as a side effect, the file is created
(if it does not exist).}

\item{several.ok}{(logical)
Specify if \code{type} should be allowed to have more than one element.
Defaults to \code{FALSE}.
See also \code{\link[base:match.arg]{base::match.arg()}}}
}
\value{
(character) Path to file.
}
\description{
Construct path to file of certain type snippets:
}
\details{
\itemize{
\item \code{get_path_to_snippet_file()}    in any folder
\item \code{get_path_to_rs_snippet_file()} in RS snippets folder
}
}
\concept{snippet files and dirs}
