% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get_rs_snippets_dir}
\alias{get_rs_snippets_dir}
\title{Get path to directory for RStudio snippets}
\usage{
get_rs_snippets_dir()
}
\value{
(string) Directory name.
}
\description{
Get path to directory for RStudio snippets
}
\examples{
get_rs_snippets_dir()
}
