% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets--install.R
\name{install-snippets}
\alias{install-snippets}
\alias{install_snippets_from_package}
\alias{install_snippets_from_dir}
\title{Install snippets}
\usage{
install_snippets_from_package(
  package = "snippets",
  type = get_default_snippet_types(),
  subdir = "",
  backup = TRUE
)

install_snippets_from_dir(
  from_dir = ".",
  type = get_default_snippet_types(),
  backup = TRUE
)
}
\arguments{
\item{package}{(character) The name of R package.}

\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{subdir}{(character) The sub-directory with replacement file(s).}

\item{backup}{(logical) Indication if a back-up copy should be created.}

\item{from_dir}{(character) The directory with replacement file.}
}
\description{
Replace current file with snippets with the other file.
}
\examples{
if (FALSE) {

# Replace your R and Markdown snippets with those in package "snippets":
install_snippets_from_package("snippets", type = "r",        backup = TRUE)
install_snippets_from_package("snippets", type = "markdown", backup = TRUE)

# Check if back-up copies exist:
list_snippet_file_backups(type = "r")
list_snippet_file_backups(type = "markdown")

}
}
\concept{install snippets}
