% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets--files-and-dirs.R
\name{make_snippet_filename}
\alias{make_snippet_filename}
\title{Make a filename for snippets}
\usage{
make_snippet_filename(type = get_default_snippet_types(), several.ok = FALSE)
}
\arguments{
\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{several.ok}{(logical)
Specify if \code{type} should be allowed to have more than one element.
Defaults to \code{FALSE}.
See also \code{\link[base:match.arg]{base::match.arg()}}}
}
\value{
String with a filename.
}
\description{
Create a string with a filename that is appropriate for certain type of
RStudio snippets.
}
\examples{
# By default, filename for R snippets is created.
make_snippet_filename()

make_snippet_filename("markdown")

make_snippet_filename(c("r", "markdown"), several.ok = TRUE)
}
\concept{snippet files and dirs}
