% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{match_snippet_type}
\alias{match_snippet_type}
\title{Return correct snippet type}
\usage{
match_snippet_type(type = get_default_snippet_types(), several.ok = FALSE)
}
\arguments{
\item{type}{A character vector of candidate values.
One of \code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"},
\code{"java"}, \code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
The default is \code{"r"}.}

\item{several.ok}{(logical) Specify if \code{type} should be allowed to have more
than one element. Default is \code{FALSE}.}
}
\value{
(sting) Correct snippet type in lower case. By default returns \code{"r"}.
}
\description{
Return correct snippet type
}
\examples{
match_snippet_type()

match_snippet_type("r")

match_snippet_type("m")
}
\seealso{
\code{\link[base:match.arg]{base::match.arg()}}
}
