% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{read_snippets}
\alias{read_snippets}
\title{Read snippets into data frame}
\usage{
read_snippets(input)
}
\arguments{
\item{input}{(character)
Either a file name or a character vector with with a snippet.}
}
\value{
File/Text with snippets parsed into data frame (\code{tibble}).
}
\description{
\lifecycle{experimental}
Read snippets from file or character vector into data frame
}
\examples{

library(snippets)

file1 <- system.file("test1/html.snippets", package = "snippets")
read_snippets(file1)

file2 <- system.file("test2/html.snippets", package = "snippets")
read_snippets(file2)

file3 <- path_to_snippets_files_of_pkg("snippets")[1]
read_snippets(file3)


text <- (
'
snippet space
	&${1:nbsp};${0}

snippet nbsp
	&${1:nbsp};${0}
')

read_snippets(text)

}
