% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{replace_snippets_file}
\alias{replace_snippets_file}
\title{Replace snippets file}
\usage{
replace_snippets_file(
  type = get_default_snippet_types(),
  from_dir = get_pkg_snippets_dir("custom"),
  backup = TRUE
)
}
\arguments{
\item{type}{The type of RStudio snippet.}

\item{from_dir}{The directory with replacement file.}

\item{backup}{(logical) Indication if a backup copy should be created.}
}
\description{
Replace snippets file
}
\examples{
if (FALSE) {

# backup_rs_snippets("r")
# backup_rs_snippets("markdown")

# FIXME:
# merge_snippets("r",        in_dir = "snippets/")
# merge_snippets("markdown", in_dir = "snippets/")
#
# replace_snippets_file("r",        from_dir = "snippets/")
# replace_snippets_file("markdown", from_dir = "snippets/")
#
# merge_snippets("r",        in_dir = "snippets/", rm = "-VG-snippets")
# merge_snippets("markdown", in_dir = "snippets/", rm = "-VG-snippets")

}
}
