% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets--files-and-dirs.R
\name{snippets_file_exists}
\alias{snippets_file_exists}
\title{Does the file with snippets exist?}
\usage{
snippets_file_exists(type, rstudio_version = "auto")
}
\arguments{
\item{type}{(character)
A character vector of snippet types. Currently allowed values are
\code{"r"}, \code{"markdown"}, \code{"c_cpp"}, \code{"css"}, \code{"html"}, \code{"java"},
\code{"javascript"}, \code{"python"}, \code{"sql"}, \verb{"stan}", \code{"tex"}.
May be unambiguously truncated.
Defaults to \code{"r"}.}

\item{rstudio_version}{Numeric version of RStudio (the output is different
for RStudio < 1.3 and 1.3 or newer series). If \code{"auto"}, the version
of RStudio will be determined automatically.}
}
\value{
Returns \code{TRUE} if file exists and \code{FALSE} otherwise.
}
\description{
Does the file with certain types of snippets exist in RStudio snippets directory?
}
\examples{
\dontrun{\donttest{
snippets_file_exists("r")
snippets_file_exists("markdown")
}}
}
\concept{snippet files and dirs}
